% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_run_hoooks.R
\name{CheckpointSaverHook}
\alias{CheckpointSaverHook}
\title{Saves checkpoints every N steps or seconds.}
\usage{
CheckpointSaverHook(checkpoint_dir, save_secs = NULL, save_steps = NULL,
  saver = NULL, checkpoint_basename = "model.ckpt", scaffold = NULL,
  listeners = NULL)
}
\arguments{
\item{checkpoint_dir}{\code{str}, base directory for the checkpoint files.}

\item{save_secs}{\code{int}, save every N secs.}

\item{save_steps}{\code{int}, save every N steps.}

\item{saver}{\code{Saver} object, used for saving.}

\item{checkpoint_basename}{\code{str}, base name for the checkpoint files.}

\item{scaffold}{\code{Scaffold}, use to get saver object.}

\item{listeners}{List of \code{CheckpointSaverListener} subclass instances.
Used for callbacks that run immediately after the corresponding CheckpointSaverHook callbacks,
only in steps where the CheckpointSaverHook was triggered.}
}
\description{
Saves checkpoints every N steps or seconds.
}
\seealso{
Other session_run_hook wrappers: \code{\link{GlobalStepWaiterHook}},
  \code{\link{LoggingTensorHook}},
  \code{\link{NanTensorHook}},
  \code{\link{StepCounterHook}},
  \code{\link{StopAtStepHook}},
  \code{\link{SummarySaverHook}}
}
