% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_custom_model.R
\name{coef.tf_custom_model}
\alias{coef.tf_custom_model}
\title{Get the list coefficients or variables from this model's checkpoint.}
\usage{
\method{coef}{tf_custom_model}(object)
}
\description{
Get the list coefficients or variables from this model's checkpoint.
}
\seealso{
Other custom estimator methods: \code{\link{estimator_spec}},
  \code{\link{estimator}},
  \code{\link{evaluate.tf_custom_model}},
  \code{\link{export_savedmodel.tf_custom_model}},
  \code{\link{predict.tf_custom_model}},
  \code{\link{train.tf_custom_model}}
}
