% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_estimator.R
\name{coef.tf_estimator}
\alias{coef.tf_estimator}
\title{Get the list coefficients or variables from this model's checkpoint.}
\usage{
\method{coef}{tf_estimator}(object)
}
\description{
Get the list coefficients or variables from this model's checkpoint.
}
\seealso{
Other custom estimator methods: \code{\link{estimator_spec}},
  \code{\link{estimator}},
  \code{\link{evaluate.tf_estimator}},
  \code{\link{export_savedmodel.tf_estimator}},
  \code{\link{predict.tf_estimator}},
  \code{\link{train.tf_estimator}}
}
