% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyselect.R
\name{column-scope}
\alias{column-scope}
\alias{set_columns}
\alias{with_columns}
\alias{scoped_columns}
\title{Establish a Feature Columns Selection Scope}
\usage{
set_columns(columns)

with_columns(columns, expr)

scoped_columns(columns)
}
\arguments{
\item{columns}{Either a named \R object (whose names will be
used to provide a selection context), or a character vector
of such names.}

\item{expr}{An \R expression, to be evaluated with the selection
context active.}
}
\description{
This helper function provides a set of names to be
used by \code{tidyselect} helpers in e.g. \code{\link[=feature_columns]{feature_columns()}}.
}
