% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_columns.R
\name{column_base}
\alias{column_base}
\title{Base Documentation for Feature Column Constructors}
\arguments{
\item{...}{Expression(s) identifying input feature(s). Used as the column
name and the dictionary key for feature parsing configs, feature tensors,
and feature columns.}
}
\description{
Base Documentation for Feature Column Constructors
}
