% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_columns.R
\name{column_one_hot}
\alias{column_one_hot}
\title{Creates an \code{_OneHotColumn} for a one-hot or multi-hot repr in a DNN.}
\usage{
column_one_hot(sparse_id_column)
}
\arguments{
\item{sparse_id_column}{A _SparseColumn which is created by \code{sparse_column_with_*} or crossed_column functions. Note that \code{combiner} defined in \code{sparse_id_column} is ignored.}
}
\value{
An _OneHotColumn.
}
\description{
Creates an \code{_OneHotColumn} for a one-hot or multi-hot repr in a DNN.
}
