% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_columns.R
\name{column_sparse_weighted}
\alias{column_sparse_weighted}
\title{Creates a _SparseColumn by combining sparse_id_column with a weight column.}
\usage{
column_sparse_weighted(sparse_id_column, weight_column_name,
  dtype = tf$float32)
}
\arguments{
\item{sparse_id_column}{A \code{_SparseColumn} which is created by \code{column_sparse_with_*} functions.}

\item{weight_column_name}{A string defining a sparse column name which represents weight or value of the corresponding sparse id feature.}

\item{dtype}{Type of weights, such as \code{tf.float32}. Only floating and integer weights are supported. Returns: A _WeightedSparseColumn composed of two sparse features: one represents id, the other represents weight (value) of the id feature in that example.}
}
\value{
A _WeightedSparseColumn composed of two sparse features: one represents id, the other represents weight (value) of the id feature in that example. Raises: ValueError: if dtype is not convertible to float.
}
\description{
Example:
}
\details{
\preformatted{sparse_feature = sparse_column_with_hash_bucket(column_name="sparse_col", hash_bucket_size=1000)
weighted_feature = weighted_sparse_column(sparse_id_column=sparse_feature, weight_column_name="weights_col")
}

This configuration assumes that input dictionary of model contains the following two items:
\itemize{
\item (key="sparse_col", value=sparse_tensor) where sparse_tensor is a SparseTensor.
\item (key="weights_col", value=weights_tensor) where weights_tensor is a SparseTensor.
Following are assumed to be true:
\item sparse_tensor.indices = weights_tensor.indices
\item sparse_tensor.dense_shape = weights_tensor.dense_shape
}
}
\section{Raises}{

ValueError: if dtype is not convertible to float.
}

\seealso{
Other feature_column wrappers: \code{\link{column_bucketized}},
  \code{\link{column_crossed}},
  \code{\link{column_embedding}},
  \code{\link{column_one_hot}},
  \code{\link{column_real_valued}},
  \code{\link{column_sparse_with_hash_bucket}},
  \code{\link{column_sparse_with_keys}}
}
