% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_columns.R
\name{column_with_hash_bucket}
\alias{column_with_hash_bucket}
\title{Creates a _SparseColumn with hashed bucket configuration.}
\usage{
column_with_hash_bucket(column_name, hash_bucket_size, combiner = "sum",
  dtype = tf$string)
}
\arguments{
\item{column_name}{A string defining sparse column name.}

\item{hash_bucket_size}{An int that is > 1. The number of buckets.}

\item{combiner}{A string specifying how to reduce if the sparse column is multivalent.
Currently "mean", "sqrtn" and "sum" are supported, with "sum" the default.
"sqrtn" often achieves good accuracy, in particular with bag-of-words columns.
\itemize{
\item "sum": do not normalize features in the column
\item "mean": do l1 normalization on features in the column
\item "sqrtn": do l2 normalization on features in the column
}

For more information: \code{tf.embedding_lookup_sparse}.}

\item{dtype}{The type of features. Only string and integer types are supported.}
}
\description{
Creates a _SparseColumn with hashed bucket configuration.
}
