% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_constructors.R
\name{construct_feature_columns}
\alias{construct_feature_columns}
\title{Construct column placeholders from vectors in an R object}
\usage{
construct_feature_columns(x, columns)
}
\description{
Construct column placeholders from vectors in an R object
}
