% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnn_estimators.R
\name{dynamic_rnn}
\alias{dynamic_rnn}
\title{Dynamic RNN Estimator}
\usage{
dynamic_rnn(problem_type, prediction_type, sequence_feature_columns,
  context_feature_columns = NULL, num_classes = NULL, num_units = NULL,
  cell_type = "basic_rnn", optimizer = "SGD", learning_rate = 0.1,
  predict_probabilities = F, momentum = NULL, gradient_clipping_norm = 5,
  dropout_keep_probabilities = NULL, feature_engineering_fn = NULL,
  model_dir = NULL, config = NULL)
}
\description{
Dynamic RNN Estimator
}
\seealso{
Other canned estimators: \code{\link{linear_classifier}},
  \code{\link{linear_dnn_combined_classifier}},
  \code{\link{linear_dnn_combined_regressor}},
  \code{\link{linear_regressor}},
  \code{\link{state_saving_rnn}}
}
