% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_custom_model.R
\name{estimator_spec}
\alias{estimator_spec}
\title{Ops and objects returned from a \code{model_fn} and passed to \code{Estimator}.}
\usage{
estimator_spec(mode, predictions = NULL, loss = NULL, train_op = NULL,
  eval_metric_ops = NULL, export_outputs = NULL,
  training_chief_hooks = NULL, training_hooks = NULL, scaffold = NULL)
}
\description{
\code{EstimatorSpec} fully defines the model to be run by \code{Estimator}.
}
\seealso{
Other custom estimator methods: \code{\link{coef.tf_custom_model}},
  \code{\link{estimator}},
  \code{\link{evaluate.tf_custom_model}},
  \code{\link{export_savedmodel.tf_custom_model}},
  \code{\link{predict.tf_custom_model}},
  \code{\link{train.tf_custom_model}}
}
