% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_run_hooks.R
\docType{class}
\name{EstimatorSessionRunHook}
\alias{EstimatorSessionRunHook}
\title{TensorFlow session run hook used in estimators}
\format{An \link{R6Class} generator object.}
\usage{
EstimatorSessionRunHook
}
\description{
This is the base R6 class used for custom session run hooks, which can be
used to monitor estimators while they are trained by TensorFlow. This class
itself does not provide any hooks; users should define their own R6 classes
extending this class and override the methods as needed.
}
\section{Methods}{

\describe{
\item{\code{begin()}}{Called once before using the session.}
\item{\code{after_create_session(session, coord)}}{Called when new TensorFlow session is created.}
\item{\code{before_run(run_context)}}{Called before each call to run().}
\item{\code{after_run(run_context, run_values)}}{Called after each call to run().}
\item{\code{end(session)}}{Called at the end of session.}
}
}

\examples{
library(tfestimators)

CustomSessionRunHook <- R6::R6Class(
  "CustomSessionRunHook",
  inherit = EstimatorSessionRunHook,
  public = list(
    end = function(session) {
      cat("Running custom session run hook at the end of a session")
    })
 )
custom_hook <- CustomSessionRunHook$new()
 
}
\seealso{
Other session_run_hook wrappers: \code{\link{hook_checkpoint_saver}},
  \code{\link{hook_global_step_waiter}},
  \code{\link{hook_logging_tensor}},
  \code{\link{hook_nan_tensor}},
  \code{\link{hook_step_counter}},
  \code{\link{hook_stop_at_step}},
  \code{\link{hook_summary_saver}}
}
\keyword{datasets}
