% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{evaluate.tf_experiment}
\alias{evaluate.tf_experiment}
\title{Evaluate on the evaluation data.}
\usage{
\method{evaluate}{tf_experiment}(object, delay_secs = NULL, ...)
}
\arguments{
\item{delay_secs}{Start evaluating after this many seconds. If \code{NULL},
defaults to using \code{self._eval_delays_secs}.}
}
\value{
The result of the \code{evaluate} call to the \code{Estimator}.
}
\description{
Runs evaluation on the evaluation data and returns the result. Runs for
\code{self._eval_steps} steps, or if it's \code{NULL}, then run until input is
exhausted or another exception is raised. Start the evaluation after
\code{delay_secs} seconds, or if it's \code{NULL}, defaults to using
\code{self._eval_delay_secs} seconds.
}
\seealso{
Other experiment methods: \code{\link{experiment.tf_estimator}},
  \code{\link{train.tf_experiment}},
  \code{\link{train_and_evaluate.tf_experiment}}
}
