% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R
\name{experiment}
\alias{experiment}
\title{Construct an Experiment}
\usage{
experiment(object, ...)
}
\arguments{
\item{object}{An \R object.}

\item{...}{Optional arguments passed on to implementing methods.}
}
\description{
Construct an experiment object.
}
