% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_methods.R
\name{export_savedmodel}
\alias{export_savedmodel}
\title{Export a Saved Model}
\usage{
export_savedmodel(object, ...)
}
\arguments{
\item{object}{An \R object.}

\item{...}{Optional arguments passed on to implementing methods.}
}
\description{
Serialize a model to disk.
}
