% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_columns.R
\name{feature_columns.default}
\alias{feature_columns.default}
\title{Construct column placeholders from vectors in an R object}
\usage{
\method{feature_columns}{default}(x, columns)
}
\description{
Construct column placeholders from vectors in an R object
}
