% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_columns.R
\name{feature_columns}
\alias{feature_columns}
\title{Define feature columns}
\usage{
feature_columns(.data, ...)
}
\arguments{
\item{.data}{Data frame from which to derive features. Note that this need only be
an exemplar of your data (names, types, factor levels, etc.) rather than all of the
data that will be used for training.}

\item{...}{One or more feature column definitions}
}
\description{
Define feature columns
}
