% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_latest_checkpoint}
\alias{get_latest_checkpoint}
\title{Function to get the latest checkpoint in a checkpoint directory}
\usage{
get_latest_checkpoint(checkpoint_dir, ...)
}
\description{
Function to get the latest checkpoint in a checkpoint directory
}
