% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_run_hooks_custom.R
\name{get_saved_history}
\alias{get_saved_history}
\title{Get Saved History}
\usage{
get_saved_history(mode_key = c("train", "eval"))
}
\arguments{
\item{mode_key}{The mode when the metrics were collected that you want to retrieve.}
}
\value{
a data.frame of metrics history
}
\description{
This function retrieves the saved metrics history via \code{\link[=hook_history_saver]{hook_history_saver()}}.
}
