% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_run_hoooks.R
\name{GlobalStepWaiterHook}
\alias{GlobalStepWaiterHook}
\title{Delay execution until global step reaches to wait_until_step.}
\usage{
GlobalStepWaiterHook(wait_until_step)
}
\arguments{
\item{wait_until_step}{an \code{int} shows until which global step should we wait.}
}
\description{
This hook delays execution until global step reaches to \code{wait_until_step}. It
is used to gradually start workers in distributed settings. One example usage
would be setting \code{wait_until_step=int(K*log(task_id+1))} assuming that
task_id=0 is the chief.
}
\seealso{
Other session_run_hook wrappers: \code{\link{CheckpointSaverHook}},
  \code{\link{LoggingTensorHook}},
  \code{\link{NanTensorHook}},
  \code{\link{StepCounterHook}},
  \code{\link{StopAtStepHook}},
  \code{\link{SummarySaverHook}}
}
