% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_run_hooks_builtin_wrappers.R
\name{hook_checkpoint_saver}
\alias{hook_checkpoint_saver}
\title{Saves Checkpoints Every N Steps or Seconds}
\usage{
hook_checkpoint_saver(checkpoint_dir, save_secs = NULL,
  save_steps = NULL, saver = NULL,
  checkpoint_basename = "model.ckpt", scaffold = NULL,
  listeners = NULL)
}
\arguments{
\item{checkpoint_dir}{The base directory for the checkpoint files.}

\item{save_secs}{An integer, indicating saving checkpoints every N secs.}

\item{save_steps}{An integer, indicating saving checkpoints every N steps.}

\item{saver}{A saver object, used for saving.}

\item{checkpoint_basename}{The base name for the checkpoint files.}

\item{scaffold}{A scaffold, used to get saver object.}

\item{listeners}{List of checkpoint saver listener subclass instances, used
for callbacks that run immediately after the corresponding
\code{hook_checkpoint_saver} callbacks, only in steps where \code{the hook_checkpoint_saver}
was triggered.}
}
\description{
Saves Checkpoints Every N Steps or Seconds
}
\seealso{
Other session_run_hook wrappers: \code{\link{hook_global_step_waiter}},
  \code{\link{hook_history_saver}},
  \code{\link{hook_logging_tensor}},
  \code{\link{hook_nan_tensor}},
  \code{\link{hook_progress_bar}},
  \code{\link{hook_step_counter}},
  \code{\link{hook_stop_at_step}},
  \code{\link{hook_summary_saver}},
  \code{\link{session_run_hook}}
}
\concept{session_run_hook wrappers}
