% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_run_hooks.R
\name{hook_global_step_waiter}
\alias{hook_global_step_waiter}
\title{Delay execution until global step reaches to wait_until_step.}
\usage{
hook_global_step_waiter(wait_until_step)
}
\arguments{
\item{wait_until_step}{an \code{int} shows until which global step should we wait.}
}
\description{
This hook delays execution until global step reaches to \code{wait_until_step}. It
is used to gradually start workers in distributed settings. One example usage
would be setting \code{wait_until_step=int(K*log(task_id+1))} assuming that
task_id=0 is the chief.
}
\seealso{
Other session_run_hook wrappers: \code{\link{EstimatorSessionRunHook}},
  \code{\link{hook_checkpoint_saver}},
  \code{\link{hook_logging_tensor}},
  \code{\link{hook_nan_tensor}},
  \code{\link{hook_step_counter}},
  \code{\link{hook_stop_at_step}},
  \code{\link{hook_summary_saver}}
}
