% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_run_hooks_custom.R
\name{hook_history_saver}
\alias{hook_history_saver}
\title{A Custom Run Hook for Saving Metrics History}
\usage{
hook_history_saver(every_n_step = 10)
}
\arguments{
\item{every_n_step}{Save the metrics every N steps}
}
\description{
This hook allows users to save the metrics history produced during training or evaluation in
a specified frequency.
}
\seealso{
Other session_run_hook wrappers: \code{\link{hook_checkpoint_saver}},
  \code{\link{hook_global_step_waiter}},
  \code{\link{hook_logging_tensor}},
  \code{\link{hook_nan_tensor}},
  \code{\link{hook_progress_bar}},
  \code{\link{hook_step_counter}},
  \code{\link{hook_stop_at_step}},
  \code{\link{hook_summary_saver}},
  \code{\link{session_run_hook}}
}
