% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_run_hooks.R
\name{hook_logging_tensor}
\alias{hook_logging_tensor}
\title{Prints the given tensors once every N local steps or once every N seconds.}
\usage{
hook_logging_tensor(tensors, every_n_iter = NULL, every_n_secs = NULL,
  formatter = NULL)
}
\arguments{
\item{tensors}{\code{dict} that maps string-valued tags to tensors/tensor names,
or \code{iterable} of tensors/tensor names.}

\item{every_n_iter}{\code{int}, print the values of \code{tensors} once every N local
steps taken on the current worker.}

\item{every_n_secs}{\code{int} or \code{float}, print the values of \code{tensors} once
every N seconds. Exactly one of \code{every_n_iter} and \code{every_n_secs} should be
provided.}

\item{formatter}{function, takes dict of \code{tag}->\code{Tensor} and returns a
string. If \code{NULL} uses default printing all tensors.}
}
\description{
The tensors will be printed to the log, with \code{INFO} severity.
}
\seealso{
Other session_run_hook wrappers: \code{\link{hook_checkpoint_saver}},
  \code{\link{hook_global_step_waiter}},
  \code{\link{hook_nan_tensor}},
  \code{\link{hook_step_counter}},
  \code{\link{hook_stop_at_step}},
  \code{\link{hook_summary_saver}}
}
