% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_run_hooks_builtin_wrappers.R
\name{hook_nan_tensor}
\alias{hook_nan_tensor}
\title{NaN Loss Monitor}
\usage{
hook_nan_tensor(loss_tensor, fail_on_nan_loss = TRUE)
}
\arguments{
\item{loss_tensor}{The loss tensor.}

\item{fail_on_nan_loss}{A boolean indicating whether to raise exception when loss is NaN.}
}
\description{
Monitors loss and stops training if loss is NaN. Can either fail with
exception or just stop training.
}
\seealso{
Other session_run_hook wrappers: \code{\link{hook_checkpoint_saver}},
  \code{\link{hook_global_step_waiter}},
  \code{\link{hook_history_saver}},
  \code{\link{hook_logging_tensor}},
  \code{\link{hook_progress_bar}},
  \code{\link{hook_step_counter}},
  \code{\link{hook_stop_at_step}},
  \code{\link{hook_summary_saver}},
  \code{\link{session_run_hook}}
}
