% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_run_hooks_custom.R
\name{hook_progress_bar}
\alias{hook_progress_bar}
\title{A Custom Run Hook to Create and Update Progress Bar During Training or Evaluation}
\usage{
hook_progress_bar()
}
\description{
This hook creates a progress bar that creates and updates the progress bar during training
or evaluation.
}
\seealso{
Other session_run_hook wrappers: \code{\link{hook_checkpoint_saver}},
  \code{\link{hook_global_step_waiter}},
  \code{\link{hook_history_saver}},
  \code{\link{hook_logging_tensor}},
  \code{\link{hook_nan_tensor}},
  \code{\link{hook_step_counter}},
  \code{\link{hook_stop_at_step}},
  \code{\link{hook_summary_saver}},
  \code{\link{session_run_hook}}
}
