% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_run_hooks_builtin_wrappers.R
\name{hook_step_counter}
\alias{hook_step_counter}
\title{Steps per Second Monitor}
\usage{
hook_step_counter(every_n_steps = 100, every_n_secs = NULL,
  output_dir = NULL, summary_writer = NULL)
}
\arguments{
\item{every_n_steps}{Run this counter every N steps}

\item{every_n_secs}{Run this counter every N seconds}

\item{output_dir}{The output directory}

\item{summary_writer}{The summary writer}
}
\description{
Steps per Second Monitor
}
\seealso{
Other session_run_hook wrappers: \code{\link{hook_checkpoint_saver}},
  \code{\link{hook_global_step_waiter}},
  \code{\link{hook_history_saver}},
  \code{\link{hook_logging_tensor}},
  \code{\link{hook_nan_tensor}},
  \code{\link{hook_progress_bar}},
  \code{\link{hook_stop_at_step}},
  \code{\link{hook_summary_saver}},
  \code{\link{session_run_hook}}
}
