% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_run_hooks_builtin_wrappers.R
\name{hook_stop_at_step}
\alias{hook_stop_at_step}
\title{Monitor to Request Stop at a Specified Step}
\usage{
hook_stop_at_step(num_steps = NULL, last_step = NULL)
}
\arguments{
\item{num_steps}{Number of steps to execute.}

\item{last_step}{Step after which to stop.}
}
\description{
Monitor to Request Stop at a Specified Step
}
\seealso{
Other session_run_hook wrappers: \code{\link{hook_checkpoint_saver}},
  \code{\link{hook_global_step_waiter}},
  \code{\link{hook_logging_tensor}},
  \code{\link{hook_nan_tensor}},
  \code{\link{hook_step_counter}},
  \code{\link{hook_summary_saver}}
}
