% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_fn.R
\name{input_fn.data.frame}
\alias{input_fn.data.frame}
\title{Construct input function from a data.frame object used to feed the estimator.}
\usage{
\method{input_fn}{data.frame}(x, features, response = NULL,
  batch_size = 10L, shuffle = TRUE, num_epochs = 1L,
  queue_capacity = 1000L, num_threads = 1L)
}
\arguments{
\item{x}{The data.frame that represents the input source}

\item{features}{The names of features to be used}

\item{response}{The response variable name to be used}

\item{batch_size}{The size of batches}

\item{shuffle}{Whether to shuffles the queue. Avoid shuffle at prediction time}

\item{num_epochs}{The number of epochs to iterate over data. If \code{NULL} will run forever.}

\item{queue_capacity}{The size of queue to accumulate.}

\item{num_threads}{The number of threads used for reading and enqueueing.
In order to have predicted and repeatable order of reading and enqueueing,
such as in prediction and evaluation mode, \code{num_threads} should be 1.}
}
\description{
Construct input function from a data.frame object used to feed the estimator.
}
\examples{
features <- c("drat", "cyl")
input_fn1 <- input_fn(mtcars, response = "mpg", features = features)

}
\seealso{
Other input function constructors: \code{\link{input_fn.default}},
  \code{\link{input_fn.list}},
  \code{\link{input_fn.matrix}}
}
