% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_fn.R
\name{input_fn.default}
\alias{input_fn.default}
\title{This function provides a skeleton of constructing a custom input function.}
\usage{
\method{input_fn}{default}(x, features, response = NULL)
}
\description{
Type \code{input_fn.default} to print out the skeleton code and create your own.
}
\seealso{
Other input function constructors: \code{\link{input_fn.data.frame}},
  \code{\link{input_fn.list}},
  \code{\link{input_fn.matrix}}
}
