% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_fn.R
\name{input_fn.list}
\alias{input_fn.list}
\title{Construct input function from a list object used to feed the estimator.}
\usage{
\method{input_fn}{list}(x, features, response)
}
\arguments{
\item{x}{The list that represents the input source}

\item{features}{The names of features to be used}

\item{response}{The response variable name to be used}
}
\description{
This is particularly useful when constructing dynamic length of inputs for
models like recurrent neural networks.
}
\examples{
input_fn(
   x = list(
     feature_names = list(
       list(list(1), list(2), list(3)),
       list(list(4), list(5), list(6))),
     response = list(
       list(1, 2, 3), list(4, 5, 6))),
   features = c("feature_names"),
   response = "response")

}
\seealso{
Other input function constructors: \code{\link{input_fn.data.frame}},
  \code{\link{input_fn.default}},
  \code{\link{input_fn.matrix}}
}
