% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_fn.R
\name{input_fn.matrix}
\alias{input_fn.matrix}
\title{Construct input function from a matrix object used to feed the estimator.}
\usage{
\method{input_fn}{matrix}(x, features, response = NULL, batch_size = 10L,
  shuffle = TRUE, num_epochs = 1L, queue_capacity = 1000L,
  num_threads = 1L)
}
\description{
This is essentially the same as \code{\link{input_fn.data.frame}}.
}
\seealso{
Other input function constructors: \code{\link{input_fn.data.frame}},
  \code{\link{input_fn.default}},
  \code{\link{input_fn.list}}
}
