% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_estimator.R
\name{keras_model_to_estimator}
\alias{keras_model_to_estimator}
\title{Keras Estimators}
\usage{
keras_model_to_estimator(keras_model = NULL, keras_model_path = NULL,
  custom_objects = NULL, model_dir = NULL, config = NULL)
}
\arguments{
\item{keras_model}{A keras model.}

\item{keras_model_path}{Directory to a keras model on disk.}

\item{custom_objects}{Dictionary for custom objects.}

\item{model_dir}{Directory to save Estimator model parameters, graph and etc.}

\item{config}{Configuration object.}
}
\description{
Create an Estimator from a compiled Keras model
}
