% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tf.R
\name{latest_checkpoint}
\alias{latest_checkpoint}
\title{Get the Latest Checkpoint in a Checkpoint Directory}
\usage{
latest_checkpoint(checkpoint_dir, ...)
}
\arguments{
\item{checkpoint_dir}{The path to the checkpoint directory.}

\item{...}{Optional arguments passed on to \code{latest_checkpoint()}.}
}
\description{
Get the Latest Checkpoint in a Checkpoint Directory
}
\seealso{
Other utility functions: \code{\link{graph_keys}}
}
