% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{learn}
\alias{learn}
\title{TF.Learn Module}
\format{An object of class \code{NULL} of length 0.}
\usage{
learn
}
\description{
\code{learn} acts as an interface to the main \href{https://www.tensorflow.org/tutorials/tflearn/}{TF.Learn}
module. Objects and functions defined within this module can be accessed
using the \code{$} function.
}
\keyword{datasets}
