% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_estimators.R
\name{linear_classification}
\alias{linear_classification}
\title{TensorFlow -- Linear Classification}
\usage{
linear_classification(recipe, run_options = NULL, ...)
}
\arguments{
\item{recipe}{A recipe. TODO: fill in the blanks.}

\item{run_options}{Optional tools used for lower-level tuning of the TensorFlow
model fitting process. See \code{\link{run_options}} for more information.}

\item{...}{Optional arguments; currently unused.}
}
\description{
Perform linear classification using TensorFlow.
}
\examples{
recipe <- simple_linear_recipe(mpg ~ drat, data = mtcars)
linear_classification(recipe = recipe)
}
