% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_dnn_combined_estimators.R
\name{linear_dnn_combined_classification}
\alias{linear_dnn_combined_classification}
\title{TensorFlow -- Linear DNN Combined Classification}
\usage{
linear_dnn_combined_classification(recipe, run.options = run_options(), ...)
}
\arguments{
\item{recipe}{A recipe. See \code{\link{tf_recipe}} for more details.}

\item{...}{Optional arguments; currently unused.}

\item{tf.options}{Optional tools used for lower-level tuning of the TensorFlow
model fitting process. See \code{\link{tf_options}} for more information.}
}
\description{
Perform Linear DNN Combined Classification using TensorFlow.
}
\examples{
recipe <- simple_linear_dnn_combined_recipe(mtcars, response = "mpg", linear.features = c("cyl"), dnn.features = c("drat"))
linear_dnn_combined_classification(recipe = recipe, dnn_hidden_units = c(10L, 10L, 10L))
}
