% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_dnn_combined_estimators.R
\name{linear_dnn_combined_classifier}
\alias{linear_dnn_combined_classifier}
\title{TensorFlow -- Linear DNN Combined Classification}
\usage{
linear_dnn_combined_classifier(linear_feature_columns, dnn_feature_columns,
  run_options = NULL, ...)
}
\arguments{
\item{run_options}{Optional tools used for lower-level tuning of the TensorFlow
model fitting process. See \code{\link{run_options}} for more information.}

\item{...}{Optional arguments; currently unused.}

\item{recipe}{A recipe. TODO: fill in the blanks.}
}
\description{
Perform Linear DNN Combined Classification using TensorFlow.
}
\examples{
recipe <- simple_linear_dnn_combined_recipe(mtcars, response = "mpg", linear_features = c("cyl"), dnn_features = c("drat"))
linear_dnn_combined_classifier(recipe = recipe, dnn_hidden_units = c(10L, 10L, 10L))
}
