% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_dnn_combined_estimators.R
\name{linear_dnn_combined_regressor}
\alias{linear_dnn_combined_regressor}
\title{TensorFlow -- Linear DNN Combined Regression}
\usage{
linear_dnn_combined_regressor(linear_feature_columns, dnn_feature_columns,
  run_options = NULL, ...)
}
\description{
Perform Linear DNN Combined Regression using TensorFlow.
}
