% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes.R
\name{linear_recipe}
\alias{linear_recipe}
\title{Construct a tflearn Recipe}
\usage{
linear_recipe(feature.columns, input.fn, ...)
}
\arguments{
\item{feature.columns}{An \R list of tensors, acting as placeholders for
input data.}

\item{input.fn}{An \R function, returning an \R list binding input data
to the aforementioned feature columns.}

\item{...}{Optional named arguments.}
}
\description{
Construct a recipe suitable for use with the higher-level
\code{tflearn} modeling routines.
}
\seealso{
Other recipes: \code{\link{simple_linear_dnn_combined_recipe}},
  \code{\link{simple_linear_recipe}}
}
