% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_estimators.R
\name{linear_regressor}
\alias{linear_regressor}
\title{TensorFlow -- Linear Regression}
\usage{
linear_regressor(feature_columns, run_options = NULL, ...)
}
\description{
Perform linear regression using TensorFlow.
}
