% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_keys.R
\name{metric_keys}
\alias{metric_keys}
\title{Enum for metric keys.}
\usage{
metric_keys()
}
\description{
Used by retrieving available metrics from canned estimators.
}
\examples{
metrics <- metric_keys()

# Get the available keys
names(metrics)

metrics$ACCURACY

}
