% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_keys.R
\name{mode_keys}
\alias{mode_keys}
\title{Standard names for model modes.}
\usage{
mode_keys()
}
\description{
The following standard keys are defined:
}
\details{
\itemize{
\item \code{TRAIN}: training mode.
\item \code{EVAL}: evaluation mode.
\item \code{PREDICT}: inference mode.
}
}
\examples{
names(mode_keys())

}
