% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tf.R
\name{model_dir}
\alias{model_dir}
\title{Model directory}
\usage{
model_dir(object, ...)
}
\arguments{
\item{object}{Model object}

\item{...}{Unused}
}
\description{
Get the directory where a model's artifacts are stored.
}
