% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_run_hoooks.R
\name{NanTensorHook}
\alias{NanTensorHook}
\title{NaN Loss monitor.}
\usage{
NanTensorHook(loss_tensor, fail_on_nan_loss = TRUE)
}
\arguments{
\item{loss_tensor}{\code{Tensor}, the loss tensor.}

\item{fail_on_nan_loss}{\code{bool}, whether to raise exception when loss is NaN.}
}
\description{
Monitors loss and stops training if loss is NaN.
Can either fail with exception or just stop training.
}
