% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_estimator.R
\name{predict.tf_estimator}
\alias{predict.tf_estimator}
\title{Generate Predictions with an Estimator}
\usage{
\method{predict}{tf_estimator}(object, input_fn, checkpoint_path = NULL,
  predict_keys = NULL, hooks = NULL, as_iterable = FALSE,
  simplify = TRUE, ...)
}
\arguments{
\item{object}{A TensorFlow estimator.}

\item{input_fn}{An input function, typically generated by the \code{\link[=input_fn]{input_fn()}}
helper function.}

\item{checkpoint_path}{The path to a specific model checkpoint to be used for
prediction. If \code{NULL} (the default), the latest checkpoint in \code{model_dir}
is used.}

\item{predict_keys}{The types of predictions that should be produced, as an
\R list. When \code{NULL} (the default), all possible predicted values will be
returned. The \code{\link[=prediction_keys]{prediction_keys()}} function provides the set of possible
prediction keys that can be supplied here.}

\item{hooks}{A list of \R functions, to be used as callbacks inside the
training loop.}

\item{as_iterable}{Boolean; should a raw Python generator be returned? When
\code{FALSE} (the default), the predicted values will be consumed from the
generator and returned as an \R object.}

\item{simplify}{A boolean value that indicates whether to simplify prediction
results automatically. If \code{TRUE}, This will automatically use a default prediction
simplify function to canned estimators that flattens the predictions to a
cleaner tibble or data.frame object with each column being the prediction results
for each \code{predict_keys} specified. For custom estimators, predictions will
not be simplified unless this parameter is a custom function with signature:
\code{function(predictions) {func(predictions)}}. If a custom function is provided
for a canned estimator, the custom function will be used instead of the default
simplify function.}

\item{...}{Optional arguments passed on to the estimator's \code{predict()}
method.}
}
\description{
Generate predicted labels / values for input data provided by \code{input_fn()}.
}
\section{Yields}{
 Evaluated values of \code{predictions} tensors.
}

\section{Raises}{
 ValueError: Could not find a trained model in model_dir.
ValueError: if batch length of predictions are not same. ValueError: If
there is a conflict between \code{predict_keys} and \code{predictions}. For example
if \code{predict_keys} is not \code{NULL} but \code{EstimatorSpec.predictions} is not a
\code{dict}.
}

\seealso{
Other custom estimator methods: \code{\link{coef.tf_estimator}},
  \code{\link{estimator_spec}}, \code{\link{estimator}},
  \code{\link{evaluate.tf_estimator}},
  \code{\link{export_savedmodel.tf_estimator}},
  \code{\link{train.tf_estimator}}
}
