% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_keys.R
\name{prediction_keys}
\alias{prediction_keys}
\title{Canonical Model Prediction Keys}
\usage{
prediction_keys()
}
\description{
The canonical set of keys used for models and estimators that provide
different types of predicted values through their \code{predict()} method.
}
\examples{
keys <- prediction_keys()

# Get the available keys
keys

# Key for retrieving probabilities from prediction values
keys$PROBABILITIES
}
\seealso{
Other estimator keys: \code{\link{metric_keys}},
  \code{\link{mode_keys}}
}
