% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_dnn_combined_estimators.R
\name{simple_linear_dnn_combined_recipe.default}
\alias{simple_linear_dnn_combined_recipe.default}
\title{#' @export
#' @family recipes
#' @examples
#' # two ways of constructing the same recipe
#' simple_linear_dnn_combined_recipe(mpg ~ cyl, data = mtcars)
#' simple_linear_dnn_combined_recipe(mtcars, response = "mpg", features = c("cyl"))
simple_linear_dnn_combined_recipe <- function(x, ...) {
  UseMethod("tf_simple_recipe")
}}
\usage{
simple_linear_dnn_combined_recipe.default(x, response, linear.features,
  dnn.features, ...)
}
\description{
#' @export
simple_linear_dnn_combined_recipe.formula <- function(x, data, ...) {
  parsed <- parse_formula(x)
  tf_simple_recipe(data, parsed$response, parsed$features)
}
}

