% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes.R
\name{simple_linear_dnn_combined_recipe}
\alias{simple_linear_dnn_combined_recipe}
\title{simple_linear_dnn_combined_recipe}
\usage{
simple_linear_dnn_combined_recipe(x, ...)
}
\description{
Simple Linear DNN Combined Recipe
}
\details{
TODO: Description
}
\examples{
## # two ways of constructing the same recipe
## simple_linear_dnn_combined_recipe(mpg ~ cyl, data = mtcars)
## simple_linear_dnn_combined_recipe(mtcars, response = "mpg", features = c("cyl"))
}
\seealso{
Other recipes: \code{\link{linear_recipe}},
  \code{\link{simple_linear_recipe}}
}
