% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnn_estimators.R
\name{state_saving_rnn}
\alias{state_saving_rnn}
\title{State Saving RNN Estimator}
\usage{
state_saving_rnn(problem_type, num_unroll, batch_size, sequence_feature_columns,
  context_feature_columns = NULL, num_classes = NULL, num_units = NULL,
  cell_type = "basic_rnn", optimizer_type = "SGD", learning_rate = 0.1,
  predict_probabilities = F, momentum = NULL, gradient_clipping_norm = 5,
  dropout_keep_probabilities = NULL, feature_engineering_fn = NULL,
  num_threads = 3L, queue_capacity = 1000L, seed = NULL,
  model_dir = NULL, config = NULL)
}
\description{
State Saving RNN Estimator
}
\seealso{
Other canned estimators: \code{\link{dnn_estimators}},
  \code{\link{dnn_linear_combined_classifier}},
  \code{\link{dnn_linear_combined_regressor}},
  \code{\link{dynamic_rnn}},
  \code{\link{linear_classifier}},
  \code{\link{linear_regressor}}
}
