% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_run_hoooks.R
\name{StepCounterHook}
\alias{StepCounterHook}
\title{Steps per second monitor.}
\usage{
StepCounterHook(every_n_steps = 100L, every_n_secs = NULL,
  output_dir = NULL, summary_writer = NULL)
}
\arguments{
\item{every_n_steps}{every_n_steps}

\item{every_n_secs}{every_n_secs}

\item{output_dir}{output_dir}

\item{summary_writer}{summary_writer}
}
\description{
Steps per second monitor.
}
