% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_run_hoooks.R
\name{StopAtStepHook}
\alias{StopAtStepHook}
\title{Monitor to request stop at a specified step.}
\usage{
StopAtStepHook(num_steps = NULL, last_step = NULL)
}
\arguments{
\item{num_steps}{Number of steps to execute.}

\item{last_step}{Step after which to stop.}
}
\description{
Monitor to request stop at a specified step.
}
