% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_estimators.R
\name{svm_classifier}
\alias{svm_classifier}
\title{TensorFlow -- Support Vector Machines}
\usage{
svm_classifier(recipe, run_options = NULL, ...)
}
\arguments{
\item{recipe}{A recipe. TODO: fill in the blanks.}

\item{run_options}{Optional tools used for lower-level tuning of the TensorFlow
model fitting process. See \code{\link{run_options}} for more information.}

\item{...}{Optional arguments; currently unused.}
}
\description{
Perform Support Vector Machines for binary classification using TensorFlow.
}
\examples{
recipe <- simple_linear_recipe(mpg ~ drat, data = mtcars)
svm_classifier(recipe = recipe)
}
