% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_config.R
\name{task_type}
\alias{task_type}
\title{Task Types}
\usage{
task_type()
}
\description{
This constant class gives the constant strings for available task types
used in \code{run_config}.
}
\examples{
task_type()$MASTER

}
\seealso{
Other run_config methods: \code{\link{run_config}}
}
