% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_utils.R
\name{tf_auto_inferred_columns}
\alias{tf_auto_inferred_columns}
\title{Construct column placeholders from vectors in an R object}
\usage{
tf_auto_inferred_columns(x, columns)
}
\description{
Construct column placeholders from vectors in an R object
}

