% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_options.R
\name{tf_options}
\alias{tf_options}
\title{Options for TensorFlow Routines}
\usage{
tf_options(optimizer = NULL, steps = 100L,
  model.dir = tf_setting("tf.model.dir", tempfile("tflearn_")))
}
\arguments{
\item{optimizer}{The optimizer to be used. When \code{NULL}, the default
optimizer associated with the active estimator will be used.}

\item{steps}{The number of steps to be used when running the associated model.}

\item{model.dir}{The location where model outputs should be written. Defaults
to a temporary directory within the \code{R} \code{\link{tempdir}}(), as
produced by \code{\link{tempfile}}().}
}
\description{
Used to control the various facets of layers constructed with the
modeling routines included in this package.
}

