% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_utils.R
\name{tf_recipe}
\alias{tf_recipe}
\title{Construct a tflearn Recipe}
\usage{
tf_recipe(feature.columns, input.fn)
}
\arguments{
\item{feature.columns}{An \R list of tensors, acting as placeholders for
input data.}

\item{input.fn}{An \R function, returning an \R list binding input data
to the aforementioned feature columns.}
}
\description{
Construct a recipe suitable for use with the higher-level
\code{tflearn} modeling routines.
}
\seealso{
Other recipes: \code{\link{tf_simple_recipe}}
}

