% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{tf_save}
\alias{tf_save}
\title{Save TensorFlow Model}
\usage{
tf_save(model, path, spec)
}
\arguments{
\item{model}{A TensorFlow model. For instance, as generated by \code{linear_regressor()}.}

\item{path}{The destination path where the model will be saved.}

\item{spec}{The model specification, each model has a default specification;
but it can be further defined specific functions like \code{tf_save_regressor_spec}.}
}
\description{
Saves a TensorFlow model with the goal of deploying and reusing across systems.
}
\seealso{
Other save: \code{\link{tf_save_spec_from_model}}
}
