% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_utils.R
\name{tf_save_classifier_spec}
\alias{tf_save_classifier_spec}
\title{Create Save-Specification for Classifiers}
\usage{
tf_save_classifier_spec(feature_columns, label_key = "label",
  label_dtype = tf$float32, label_default = NULL, weight_column = NULL)
}
\arguments{
\item{feature_columns}{An iterable containing all feature columns. All items
should be instances of classes derived from \code{_FeatureColumn}.}

\item{label_key}{A string identifying the label. It means \code{tf$Example} stores
labels with this key.}

\item{label_dtype}{A \code{tf$dtype} identifies the type of labels. By default it
is \code{tf$int64}. If user defines a \code{label_vocabulary}, this should be set as
\code{tf$string}. \code{tf$float32} labels are only supported for binary
classification.}

\item{label_default}{used as label if label_key does not exist in given
\code{tf$Example}. An example usage: let's say \code{label_key} is 'clicked' and
\code{tf$Example} contains clicked data only for positive examples in following
format \code{key:clicked, value:1}. This means that if there is no data with key
'clicked' it should count as negative example by setting
\code{label_deafault=0}. Type of this value should be compatible with
\code{label_dtype}.}

\item{weight_column}{A string or a numeric column created by
\code{\link[=column_numeric]{column_numeric()}} defining feature column representing
weights. It is used to down weight or boost examples during training. It
will be multiplied by the loss of the example. If it is a string, it is
used as a key to fetch weight tensor from the \code{features}. If it is a
numeric column, raw tensor is fetched by key \code{weight_column$key}, then
\code{weight_column$normalizer_fn} is applied on it to get weight tensor.}
}
\description{
Creates specification to save classification models.
}
\seealso{
Other parsing utilities: \code{\link{classifier_parse_example_spec}},
  \code{\link{regressor_parse_example_spec}},
  \code{\link{tf_save_regressor_spec}}
}
