% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_utils.R
\name{tf_save_regressor_spec}
\alias{tf_save_regressor_spec}
\title{Creates Save-Specification for Regressors}
\usage{
tf_save_regressor_spec(feature_columns, label_key = "label",
  label_dtype = tf$float32, label_default = NULL, label_dimension = 1,
  weight_column = NULL)
}
\arguments{
\item{feature_columns}{An iterable containing all feature columns. All items
should be instances of classes derived from \code{_FeatureColumn}.}

\item{label_key}{A string identifying the label. It means \code{tf$Example} stores
labels with this key.}

\item{label_dtype}{A \code{tf$dtype} identifies the type of labels. By default it
is \code{tf$float32}.}

\item{label_default}{used as label if label_key does not exist in given
\code{tf$Example}. By default default_value is none, which means
\code{tf$parse_example} will error out if there is any missing label.}

\item{label_dimension}{Number of regression targets per example. This is the
size of the last dimension of the labels and logits \code{Tensor} objects
(typically, these have shape \code{[batch_size, label_dimension]}).}

\item{weight_column}{A string or a \code{_NumericColumn} created by
\code{column_numeric} defining feature column representing
weights. It is used to down weight or boost examples during training. It
will be multiplied by the loss of the example. If it is a string, it is
used as a key to fetch weight tensor from the \code{features}. If it is a
\code{_NumericColumn}, raw tensor is fetched by key \code{weight_column$key}, then
\code{weight_column$normalizer_fn} is applied on it to get weight tensor.}
}
\description{
Creates specification to save regression models.
}
\seealso{
Other parsing utilities: \code{\link{classifier_parse_example_spec}},
  \code{\link{regressor_parse_example_spec}},
  \code{\link{tf_save_classifier_spec}}
}
