% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{tf_save_spec_from_model}
\alias{tf_save_spec_from_model}
\title{Create Save-Specification from Model}
\usage{
tf_save_spec_from_model(model)
}
\arguments{
\item{model}{A TensorFlow model. For instance, as generated by \code{linear_regressor()}.}
}
\description{
Creates default specification for a given model.
}
\seealso{
Other save: \code{\link{tf_save}}
}
