% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_utils.R
\name{tf_simple_recipe}
\alias{tf_simple_recipe}
\title{Construct a Simple tflearn Recipe}
\usage{
tf_simple_recipe(x, ...)
}
\arguments{
\item{x}{An \R object; typically a \code{data.frame} or a \code{formula}. See
examples for usage.}

\item{...}{Optional arguments passed to implementing methods.}
}
\description{
Construct a simple recipe suitable for use with the higher-level
\code{tflearn} modeling routines. This can be used to directly model a
response variable as a function of a set of untransformed features in a
dataset.
}
\examples{
# two ways of constructing the same recipe
tf_simple_recipe(mpg ~ cyl, data = mtcars)
tf_simple_recipe(mtcars, response = "mpg", features = c("cyl"))
}
\seealso{
Other recipes: \code{\link{tf_recipe}}
}

