% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_estimator.R
\name{train-evaluate-predict}
\alias{train-evaluate-predict}
\title{Base Documentation for train, evaluate, and predict.}
\arguments{
\item{input_fn}{An input function, typically generated by the \code{\link[=input_fn]{input_fn()}}
helper function.}

\item{hooks}{A list of \R functions, to be used as callbacks inside the
training loop. By default, \code{hook_history_saver(every_n_step = 10)} and
\code{hook_progress_bar()} will be attached if not provided to save the metrics
history and create the progress bar.}

\item{checkpoint_path}{The path to a specific model checkpoint to be used for
prediction. If \code{NULL} (the default), the latest checkpoint in \code{model_dir}
is used.}
}
\description{
Base Documentation for train, evaluate, and predict.
}
