% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{train.tf_experiment}
\alias{train.tf_experiment}
\title{Fit the estimator using the training data.}
\usage{
\method{train}{tf_experiment}(object, delay_secs = NULL, ...)
}
\arguments{
\item{delay_secs}{Start training after this many seconds.}
}
\value{
The trained estimator.
}
\description{
Train the estimator for \code{self._train_steps} steps, after waiting for
\code{delay_secs} seconds. If \code{self._train_steps} is \code{NULL}, train forever.
}
\seealso{
Other experiment methods: \code{\link{evaluate.tf_experiment}},
  \code{\link{experiment.tf_estimator}},
  \code{\link{train_and_evaluate.tf_experiment}}
}
