% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_estimator.R
\name{variable_names_values}
\alias{variable_names_values}
\alias{variable_names}
\alias{variable_value}
\title{Get variable names and values associated with an estimator}
\usage{
variable_names(object)

variable_value(object, variable = NULL)
}
\arguments{
\item{object}{A trained estimator model.}

\item{variable}{(Optional) Names of variables to extract as a character vector. If not specified, values for all variables are returned.}
}
\value{
For \code{variable_names()}, a vector of variable names. For \code{variable_values()}, a named list of variable values.
}
\description{
These helper functions extract the names and values of variables
in the graphs associated with trained estimator models.
}
