% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_estimator.R
\name{coef.tf_estimator}
\alias{coef.tf_estimator}
\title{Extract Model Coefficients}
\usage{
\method{coef}{tf_estimator}(object, ...)
}
\arguments{
\item{object}{An estimator.}

\item{...}{Optional arguments; currently unused.}
}
\value{
A named list of variables, such as each hidden layer's biases and
weights matrices (if using any of the DNN estimators), global step, etc.
}
\description{
Extract Model Coefficients
}
\seealso{
Other custom estimator methods: \code{\link{estimator_spec}},
  \code{\link{estimator}},
  \code{\link{evaluate.tf_estimator}},
  \code{\link{export_savedmodel.tf_estimator}},
  \code{\link{predict.tf_estimator}},
  \code{\link{train.tf_estimator}}
}
