% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_columns.R
\name{column_bucketized}
\alias{column_bucketized}
\title{Creates a _BucketizedColumn for discretizing dense input.}
\usage{
column_bucketized(source_column, boundaries)
}
\arguments{
\item{source_column}{A _RealValuedColumn defining dense column.}

\item{boundaries}{A list or list of floats specifying the boundaries. It has to be sorted.}
}
\value{
A _BucketizedColumn.
}
\description{
Creates a _BucketizedColumn for discretizing dense input.
}
\section{Raises}{

ValueError: if 'boundaries' is empty or not sorted.
}

\seealso{
Other feature_column wrappers: \code{\link{column_crossed}},
  \code{\link{column_embedding}},
  \code{\link{column_one_hot}},
  \code{\link{column_real_valued}},
  \code{\link{column_sparse_weighted}},
  \code{\link{column_sparse_with_hash_bucket}},
  \code{\link{column_sparse_with_keys}}
}
