% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_columns.R
\name{column_bucketized}
\alias{column_bucketized}
\title{Represents discretized dense input.}
\usage{
column_bucketized(source_column, boundaries)
}
\arguments{
\item{source_column}{A one-dimensional dense column which is generated with \code{numeric_column}.}

\item{boundaries}{A sorted list or list of floats specifying the boundaries.}
}
\value{
A \code{_BucketizedColumn}.
}
\description{
Buckets include the left boundary, and exclude the right boundary. Namely,
\code{boundaries=[0., 1., 2.]} generates buckets \code{(-inf, 0.)}, \code{[0., 1.)},
\code{[1., 2.)}, and \code{[2., +inf)}. For example, if the inputs are \code{boundaries} = [0, 10, 100]
input tensor = [[-5, 10000] [150, 10] [5, 100]]
then the output will be output = [[0, 3] [3, 2] [1, 3]]\preformatted{
[0., 1., 2.]: R:0.,%201.,%202.
[0., 1.)`,
`[1., 2.)`, and `[2., +inf)`. For example, if the inputs are `boundaries` = [0, 10, 100]: R:0.,%201.)%60,%0A%60[1.,%202.)%60,%20and%20%60[2.,%20+inf)%60.%20For%20example,%20if%20the%20inputs%20are%20%60boundaries%60%20=%20[0,%2010,%20100
[[-5, 10000]: R:[-5,%2010000
[150, 10]: R:150,%2010
[5, 100]: R:5,%20100
[[0, 3]: R:[0,%203
[3, 2]: R:3,%202
[1, 3]: R:1,%203
}
}
\section{Raises}{

ValueError: If \code{source_column} is not a numeric column, or if it is not one-dimensional.
ValueError: If \code{boundaries} is not a sorted list or list.
}

\seealso{
Other feature_column wrappers: \code{\link{column_categorical_with_hash_bucket}},
  \code{\link{column_categorical_with_identity}},
  \code{\link{column_categorical_with_vocabulary_file}},
  \code{\link{column_categorical_with_vocabulary_list}},
  \code{\link{column_crossed}},
  \code{\link{column_embedding}},
  \code{\link{column_numeric}},
  \code{\link{column_weighted_categorical}}
}
