% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_columns.R
\name{column_categorical_with_vocabulary_file}
\alias{column_categorical_with_vocabulary_file}
\title{A \code{_CategoricalColumn} with a vocabulary file.}
\usage{
column_categorical_with_vocabulary_file(key, vocabulary_file, vocabulary_size,
  num_oov_buckets = 0L, default_value = NULL, dtype = tf$string)
}
\arguments{
\item{key}{A unique string identifying the input feature. It is used as the column name and the dictionary key for feature parsing configs, feature \code{Tensor} objects, and feature columns.}

\item{vocabulary_file}{The vocabulary file name.}

\item{vocabulary_size}{Number of the elements in the vocabulary. This must be no greater than length of \code{vocabulary_file}, if less than length, later values are ignored.}

\item{num_oov_buckets}{Non-negative integer, the number of out-of-vocabulary buckets. All out-of-vocabulary inputs will be assigned IDs in the range \code{[vocabulary_size, vocabulary_size+num_oov_buckets)} based on a hash of the input value. A positive \code{num_oov_buckets} can not be specified with \code{default_value}.}

\item{default_value}{The integer ID value to return for out-of-vocabulary feature values, defaults to \code{-1}. This can not be specified with a positive \code{num_oov_buckets}.}

\item{dtype}{The type of features. Only string and integer types are supported.}
}
\value{
A \code{_CategoricalColumn} with a vocabulary file.
}
\description{
Use this when your inputs are in string or integer format, and you have a
vocabulary file that maps each value to an integer ID. By default,
out-of-vocabulary values are ignored. Use either (but not both) of
\code{num_oov_buckets} and \code{default_value} to specify how to include
out-of-vocabulary values. For input dictionary \code{features}, \code{features[key]} is either \code{Tensor} or
\code{SparseTensor}. If \code{Tensor}, missing values can be represented by \code{-1} for int
and \code{''} for string. Note that these values are independent of the
\code{default_value} argument. Example with \code{num_oov_buckets}:
File '/us/states.txt' contains 50 lines, each with a 2-character U.S. state
abbreviation. All inputs with values in that file are assigned an ID 0-49,
corresponding to its line number. All other values are hashed and assigned an
ID 50-54.
}
\section{Raises}{

ValueError: \code{vocabulary_file} is missing.
ValueError: \code{vocabulary_size} is missing or < 1.
ValueError: \code{num_oov_buckets} is not a non-negative integer.
ValueError: \code{dtype} is neither string nor integer.
}

\seealso{
Other feature_column wrappers: \code{\link{column_bucketized}},
  \code{\link{column_categorical_with_hash_bucket}},
  \code{\link{column_categorical_with_identity}},
  \code{\link{column_categorical_with_vocabulary_list}},
  \code{\link{column_crossed}},
  \code{\link{column_embedding}},
  \code{\link{column_numeric}},
  \code{\link{column_weighted_categorical}}
}
