% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_columns.R
\name{column_categorical_with_vocabulary_list}
\alias{column_categorical_with_vocabulary_list}
\title{A \code{_CategoricalColumn} with in-memory vocabulary.}
\usage{
column_categorical_with_vocabulary_list(fc, key, vocabulary_list,
  dtype = NULL, default_value = -1L)
}
\arguments{
\item{fc}{Feature columns}

\item{key}{Expression identifying input feature(s). Used as the column name
and the dictionary key for feature parsing configs, feature \code{Tensor}
objects, and feature columns.}

\item{vocabulary_list}{An ordered iterable defining the vocabulary.
Each feature is mapped to the index of its value (if present) in \code{vocabulary_list}.
Must be castable to \code{dtype}.}

\item{dtype}{The type of features. Only string and integer types are supported.
If \code{NULL}, it will be inferred from \code{vocabulary_list}.}

\item{default_value}{The value to use for values not in \code{vocabulary_list}.}
}
\value{
A \code{_CategoricalColumn} with in-memory vocabulary.
}
\description{
Use this when your inputs are in string or integer format, and you have an
in-memory vocabulary mapping each value to an integer ID. By default,
out-of-vocabulary values are ignored. Use \code{default_value} to specify how to
include out-of-vocabulary values. For input dictionary \code{features}, \code{features$key} is either \code{Tensor} or
\code{SparseTensor}. If \code{Tensor}, missing values can be represented by \code{-1} for int
and \code{''} for string.
}
\details{
Note that these values are independent of the
\code{default_value} argument.
}
\section{Raises}{

ValueError: if \code{vocabulary_list} is empty, or contains duplicate keys. ValueError: if \code{dtype} is not integer or string.
}

\seealso{
Other feature_column wrappers: \code{\link{column_bucketized}},
  \code{\link{column_categorical_with_hash_bucket}},
  \code{\link{column_categorical_with_identity}},
  \code{\link{column_categorical_with_vocabulary_file}},
  \code{\link{column_crossed}},
  \code{\link{column_embedding}},
  \code{\link{column_numeric}},
  \code{\link{column_weighted_categorical}}
}
