% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_columns.R
\name{column_indicator}
\alias{column_indicator}
\title{Represents multi-hot representation of given categorical column.}
\usage{
column_indicator(categorical_column)
}
\arguments{
\item{categorical_column}{A \code{_CategoricalColumn} which is created by
\code{categorical_column_with_*} or \code{crossed_column} functions.}
}
\value{
An \code{_IndicatorColumn}.
}
\description{
Used to wrap any \code{categorical_column_*} (e.g., to feed to DNN). Use
\code{embedding_column} if the inputs are sparse.
}
