% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_columns.R
\name{column_indicator}
\alias{column_indicator}
\title{Represents Multi-Hot Representation of Given Categorical Column}
\usage{
column_indicator(categorical_column)
}
\arguments{
\item{categorical_column}{A categorical column which is created by
the \code{column_categorical_with_*()} or \code{column_crossed()} functions.}
}
\value{
An indicator column.
}
\description{
Used to wrap any \code{column_categorical()*} (e.g., to feed to DNN). Use
\code{column_embedding()} if the inputs are sparse.
}
