% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_columns.R
\name{column_weighted_categorical}
\alias{column_weighted_categorical}
\title{Applies weight values to a \code{_CategoricalColumn}.}
\usage{
column_weighted_categorical(categorical_column, weight_feature_key,
  dtype = tf$float32)
}
\arguments{
\item{categorical_column}{A \code{_CategoricalColumn} created by \code{categorical_column_with_*} functions.}

\item{weight_feature_key}{String key for weight values.}

\item{dtype}{Type of weights, such as \code{tf.float32}. Only float and integer weights are supported.}
}
\value{
A \code{_CategoricalColumn} composed of two sparse features: one represents id, the other represents weight (value) of the id feature in that example.
}
\description{
Use this when each of your sparse inputs has both an ID and a value. For
example, if you're representing text documents as a collection of word
frequencies, you can provide 2 parallel sparse input features ('terms' and
'frequencies' below).
}
\section{Raises}{

ValueError: if \code{dtype} is not convertible to float.
}

\seealso{
Other feature_column wrappers: \code{\link{column_bucketized}},
  \code{\link{column_categorical_with_hash_bucket}},
  \code{\link{column_categorical_with_identity}},
  \code{\link{column_categorical_with_vocabulary_file}},
  \code{\link{column_categorical_with_vocabulary_list}},
  \code{\link{column_crossed}},
  \code{\link{column_embedding}},
  \code{\link{column_numeric}}
}
