% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_custom_model.R
\name{estimator}
\alias{estimator}
\title{Custom estimator constructor}
\usage{
estimator(model_fn, run_options = NULL, ...)
}
\description{
Custom estimator constructor
}
\seealso{
Other custom estimator: \code{\link{estimator_spec}},
  \code{\link{evaluate.tf_custom_model}},
  \code{\link{export_savedmodel.tf_custom_model}},
  \code{\link{fit.tf_custom_model}},
  \code{\link{predict.tf_custom_model}}
}
