% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_custom_model.R
\name{estimator_spec}
\alias{estimator_spec}
\title{Ops and objects returned from a \code{model_fn} and passed to \code{Estimator}.}
\usage{
estimator_spec(predictions, loss, train_op, mode)
}
\description{
\code{EstimatorSpec} fully defines the model to be run by \code{Estimator}.
}
\seealso{
Other custom estimator methods: \code{\link{estimator}},
  \code{\link{evaluate.tf_custom_model}},
  \code{\link{export_savedmodel.tf_custom_model}},
  \code{\link{fit.tf_custom_model}},
  \code{\link{predict.tf_custom_model}}
}
