% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_columns.R
\name{feature_columns}
\alias{feature_columns}
\title{Define feature columns}
\usage{
feature_columns(data, ...)
}
\arguments{
\item{data}{Exemplar of data to extract feature columns for. This need only
be a smaller sample of your input data (enough to provide column names and
types).}

\item{...}{Columns to include as features.}
}
\description{
Define feature columns
}
\details{
Columns can be specified as symbols or character vectors, and the \code{-}
operator can be used to select all but one other column.

The specified columns will receive the default feature transformation
behavior for their type (\code{column_numeric()} for numeric data,
\code{column_with_keys()} for other data). You can chain the \code{column_} family
of functions onto \code{feature_columns()} to do more customized transformations.
}
