% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_columns.R
\name{feature_columns}
\alias{feature_columns}
\title{Create list of feature columns}
\usage{
feature_columns(..., names = NULL)
}
\arguments{
\item{...}{One or more feature column definitions.}

\item{names}{Available feature names (for selection / pattern matching) as a
character vector (or R object that implements \code{names()} or \code{colnames()})}
}
\description{
Create list of feature columns
}
