% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tf.R
\name{get_latest_checkpoint}
\alias{get_latest_checkpoint}
\title{Function to get the latest checkpoint in a checkpoint directory}
\usage{
get_latest_checkpoint(checkpoint_dir, ...)
}
\arguments{
\item{checkpoint_dir}{The path to the checkpoint directory.}

\item{...}{Optional arguments passed on to \code{latest_checkpoint()}.}
}
\description{
Function to get the latest checkpoint in a checkpoint directory
}
