% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_run_hooks.R
\name{hook_checkpoint_saver}
\alias{hook_checkpoint_saver}
\title{Saves checkpoints every N steps or seconds.}
\usage{
hook_checkpoint_saver(checkpoint_dir, save_secs = NULL, save_steps = NULL,
  saver = NULL, checkpoint_basename = "model.ckpt", scaffold = NULL,
  listeners = NULL)
}
\arguments{
\item{checkpoint_dir}{\code{str}, base directory for the checkpoint files.}

\item{save_secs}{\code{int}, save every N secs.}

\item{save_steps}{\code{int}, save every N steps.}

\item{saver}{\code{Saver} object, used for saving.}

\item{checkpoint_basename}{\code{str}, base name for the checkpoint files.}

\item{scaffold}{\code{Scaffold}, use to get saver object.}

\item{listeners}{List of \code{CheckpointSaverListener} subclass instances. Used
for callbacks that run immediately after the corresponding
CheckpointSaverHook callbacks, only in steps where the CheckpointSaverHook
was triggered.}
}
\description{
Saves checkpoints every N steps or seconds.
}
\seealso{
Other session_run_hook wrappers: \code{\link{EstimatorSessionRunHook}},
  \code{\link{hook_global_step_waiter}},
  \code{\link{hook_logging_tensor}},
  \code{\link{hook_nan_tensor}},
  \code{\link{hook_step_counter}},
  \code{\link{hook_stop_at_step}},
  \code{\link{hook_summary_saver}}
}
