% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_run_hooks.R
\name{hook_step_counter}
\alias{hook_step_counter}
\title{Steps per second monitor.}
\usage{
hook_step_counter(every_n_steps = 100L, every_n_secs = NULL,
  output_dir = NULL, summary_writer = NULL)
}
\arguments{
\item{every_n_steps}{every_n_steps}

\item{every_n_secs}{every_n_secs}

\item{output_dir}{output_dir}

\item{summary_writer}{summary_writer}
}
\description{
Steps per second monitor.
}
\seealso{
Other session_run_hook wrappers: \code{\link{EstimatorSessionRunHook}},
  \code{\link{hook_checkpoint_saver}},
  \code{\link{hook_global_step_waiter}},
  \code{\link{hook_logging_tensor}},
  \code{\link{hook_nan_tensor}},
  \code{\link{hook_stop_at_step}},
  \code{\link{hook_summary_saver}}
}
