% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_run_hooks.R
\name{hook_summary_saver}
\alias{hook_summary_saver}
\title{Saves summaries every N steps.}
\usage{
hook_summary_saver(save_steps = NULL, save_secs = NULL, output_dir = NULL,
  summary_writer = NULL, scaffold = NULL, summary_op = NULL)
}
\arguments{
\item{save_steps}{\code{int}, save summaries every N steps. Exactly one of
\code{save_secs} and \code{save_steps} should be set.}

\item{save_secs}{\code{int}, save summaries every N seconds.}

\item{output_dir}{\code{string}, the directory to save the summaries to. Only used
if no \code{summary_writer} is supplied.}

\item{summary_writer}{\code{SummaryWriter}. If \code{NULL} and an \code{output_dir} was
passed, one will be created accordingly.}

\item{scaffold}{\code{Scaffold} to get summary_op if it's not provided.}

\item{summary_op}{\code{Tensor} of type \code{string} containing the serialized
\code{Summary} protocol buffer or a list of \code{Tensor}. They are most likely an
output by TF summary methods like \code{tf.summary.scalar} or
\code{tf.summary.merge_all}. It can be passed in as one tensor; if more than
one, they must be passed in as a list.}
}
\description{
Saves summaries every N steps.
}
\seealso{
Other session_run_hook wrappers: \code{\link{hook_checkpoint_saver}},
  \code{\link{hook_global_step_waiter}},
  \code{\link{hook_logging_tensor}},
  \code{\link{hook_nan_tensor}},
  \code{\link{hook_step_counter}},
  \code{\link{hook_stop_at_step}}
}
