% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_estimators.R
\name{linear_classification}
\alias{linear_classification}
\title{TensorFlow -- Linear Classification}
\usage{
linear_classification(recipe, run_options = NULL, skip_fit = FALSE, ...)
}
\arguments{
\item{recipe}{A recipe. See \code{\link{tf_recipe}} for more details.}

\item{...}{Optional arguments; currently unused.}

\item{tf.options}{Optional tools used for lower-level tuning of the TensorFlow
model fitting process. See \code{\link{tf_options}} for more information.}
}
\description{
Perform linear classification using TensorFlow.
}
\examples{
recipe <- simple_linear_recipe(mpg ~ drat, data = mtcars)
linear_classification(recipe = recipe)
}
