% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_estimators.R
\name{linear_classifier}
\alias{linear_classifier}
\title{TensorFlow -- Linear Classification}
\usage{
linear_classifier(feature_columns, n_classes = 2L, run_options = NULL, ...)
}
\description{
Perform linear classification using TensorFlow.
}
