% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_dnn_combined_estimators.R
\name{linear_dnn_combined_classifier}
\alias{linear_dnn_combined_classifier}
\title{TensorFlow -- Linear DNN Combined Classification}
\usage{
linear_dnn_combined_classifier(linear_feature_columns, dnn_feature_columns,
  run_options = NULL, ...)
}
\description{
Perform Linear DNN Combined Classification using TensorFlow.
}
