% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_dnn_combined_estimators.R
\name{linear_dnn_combined_regressor}
\alias{linear_dnn_combined_regressor}
\title{TensorFlow -- Linear DNN Combined Regression}
\usage{
linear_dnn_combined_regressor(linear_feature_columns, dnn_feature_columns,
  run_options = NULL, ...)
}
\arguments{
\item{run_options}{Optional tools used for lower-level tuning of the TensorFlow
model fitting process. See \code{\link{run_options}} for more information.}

\item{...}{Optional arguments; currently unused.}

\item{recipe}{A recipe. TODO: fill in the blanks.}
}
\description{
Perform Linear DNN Combined Regression using TensorFlow.
}
