% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_estimators.R
\name{linear_regression}
\alias{linear_regression}
\title{TensorFlow -- Linear Regression}
\usage{
linear_regression(recipe, run.options = run_options(), ...)
}
\arguments{
\item{recipe}{A recipe. See \code{\link{tf_recipe}} for more details.}

\item{...}{Optional arguments; currently unused.}

\item{tf.options}{Optional tools used for lower-level tuning of the TensorFlow
model fitting process. See \code{\link{tf_options}} for more information.}
}
\description{
Perform linear regression using TensorFlow.
}
\examples{
recipe <- simple_linear_recipe(mpg ~ drat, data = mtcars)
linear_regression(recipe = recipe)
}

