% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_estimators.R
\name{linear_regressor}
\alias{linear_regressor}
\title{TensorFlow -- Linear Regression}
\usage{
linear_regressor(feature_columns, run_options = NULL, ...)
}
\arguments{
\item{run_options}{Optional tools used for lower-level tuning of the TensorFlow
model fitting process. See \code{\link{run_options}} for more information.}

\item{...}{Optional arguments; currently unused.}

\item{recipe}{A recipe. TODO: fill in the blanks.}
}
\description{
Perform linear regression using TensorFlow.
}
