% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_keys.R
\name{metric_keys}
\alias{metric_keys}
\title{Canonical Metric Keys}
\usage{
metric_keys()
}
\description{
The canonical set of keys that can be used to access metrics from canned
estimators.
}
\examples{
\dontrun{
metrics <- metric_keys()

# Get the available keys
metrics

metrics$ACCURACY
}

}
\seealso{
Other estimator keys: 
\code{\link{mode_keys}()},
\code{\link{prediction_keys}()}
}
\concept{estimator keys}
