% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_keys.R
\name{mode_keys}
\alias{mode_keys}
\title{Canonical Mode Modes}
\usage{
mode_keys()
}
\description{
The names for different possible modes for an estimator. The following
standard keys are defined:
}
\details{
\tabular{ll}{
\code{TRAIN}   \tab Training mode. \cr
\code{EVAL}    \tab Evaluation mode. \cr
\code{PREDICT} \tab Prediction / inference mode. \cr
}
}
\examples{
modes <- mode_keys()
modes$TRAIN

}
\seealso{
Other estimator keys: \code{\link{metric_keys}},
  \code{\link{prediction_keys}}
}
