% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_keys.R
\name{prediction_keys}
\alias{prediction_keys}
\title{Enum for canonical model prediction keys.}
\usage{
prediction_keys()
}
\description{
Used by models that predict values, such as regressor models.
}
\examples{
keys <- prediction_keys()

# Get the available keys
names(keys)

# Key for retrieving probabilities from prediction values
keys$PROBABILITIES
}
