% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-tidyselect.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{contains}
\alias{select_helpers}
\alias{ends_with}
\alias{everything}
\alias{matches}
\alias{num_range}
\alias{one_of}
\alias{starts_with}
\alias{last_col}
\alias{\%>\%}
\alias{use_python}
\alias{use_virtualenv}
\alias{use_condaenv}
\alias{array_reshape}
\alias{tf}
\alias{shape}
\alias{install_tensorflow}
\alias{tf_config}
\alias{tensorboard}
\alias{evaluate}
\alias{train}
\alias{train_and_evaluate}
\alias{export_savedmodel}
\alias{flags}
\alias{flag_numeric}
\alias{flag_integer}
\alias{flag_string}
\alias{flag_boolean}
\alias{run_dir}
\alias{glimpse}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{reticulate}{\code{\link[reticulate]{use_python}}, \code{\link[reticulate]{use_virtualenv}}, \code{\link[reticulate]{use_condaenv}}, \code{\link[reticulate]{array_reshape}}}

  \item{tensorflow}{\code{\link[tensorflow]{tf}}, \code{\link[tensorflow]{shape}}, \code{\link[tensorflow]{install_tensorflow}}, \code{\link[tensorflow]{tf_config}}, \code{\link[tensorflow]{tensorboard}}, \code{\link[tensorflow]{evaluate}}, \code{\link[tensorflow]{train}}, \code{\link[tensorflow]{train_and_evaluate}}, \code{\link[tensorflow]{export_savedmodel}}}

  \item{tfruns}{\code{\link[tfruns]{flags}}, \code{\link[tfruns]{flag_numeric}}, \code{\link[tfruns]{flag_integer}}, \code{\link[tfruns]{flag_string}}, \code{\link[tfruns]{flag_boolean}}, \code{\link[tfruns]{run_dir}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}}

  \item{tidyselect}{\code{\link[tidyselect]{contains}}, \code{\link[tidyselect]{ends_with}}, \code{\link[tidyselect]{everything}}, \code{\link[tidyselect]{matches}}, \code{\link[tidyselect]{num_range}}, \code{\link[tidyselect]{one_of}}, \code{\link[tidyselect]{starts_with}}, \code{\link[tidyselect]{last_col}}}
}}

