% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{use_python}
\alias{reexports}
\alias{use_virtualenv}
\alias{reexports}
\alias{use_condaenv}
\alias{reexports}
\alias{tf}
\alias{reexports}
\alias{install_tensorflow}
\alias{reexports}
\alias{tf_config}
\alias{reexports}
\alias{tensorboard}
\alias{reexports}
\alias{evaluate}
\alias{reexports}
\alias{train}
\alias{reexports}
\alias{train_and_evaluate}
\alias{reexports}
\alias{export_savedmodel}
\alias{reexports}
\alias{flags}
\alias{reexports}
\alias{flag_numeric}
\alias{reexports}
\alias{flag_integer}
\alias{reexports}
\alias{flag_string}
\alias{reexports}
\alias{flag_boolean}
\alias{reexports}
\alias{run_dir}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{reticulate}{\code{\link[reticulate]{use_python}}, \code{\link[reticulate]{use_virtualenv}}, \code{\link[reticulate]{use_condaenv}}}

  \item{tensorflow}{\code{\link[tensorflow]{tf}}, \code{\link[tensorflow]{install_tensorflow}}, \code{\link[tensorflow]{tf_config}}, \code{\link[tensorflow]{tensorboard}}, \code{\link[tensorflow]{evaluate}}, \code{\link[tensorflow]{train}}, \code{\link[tensorflow]{train_and_evaluate}}, \code{\link[tensorflow]{export_savedmodel}}}

  \item{tfruns}{\code{\link[tfruns]{flags}}, \code{\link[tfruns]{flag_numeric}}, \code{\link[tfruns]{flag_integer}}, \code{\link[tfruns]{flag_string}}, \code{\link[tfruns]{flag_boolean}}, \code{\link[tfruns]{run_dir}}}
}}

