% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{use_python}
\alias{reexports}
\alias{use_virtualenv}
\alias{reexports}
\alias{use_condaenv}
\alias{reexports}
\alias{tf}
\alias{reexports}
\alias{install_tensorflow}
\alias{reexports}
\alias{tf_config}
\alias{reexports}
\alias{tensorboard}
\alias{reexports}
\alias{unique_dir}
\alias{reexports}
\alias{run_dir}
\alias{reexports}
\alias{starts_with}
\alias{reexports}
\alias{ends_with}
\alias{reexports}
\alias{contains}
\alias{reexports}
\alias{matches}
\alias{reexports}
\alias{everything}
\alias{reexports}
\alias{one_of}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{reticulate}{\code{\link[reticulate]{use_python}}, \code{\link[reticulate]{use_virtualenv}}, \code{\link[reticulate]{use_condaenv}}}

  \item{tensorflow}{\code{\link[tensorflow]{tf}}, \code{\link[tensorflow]{install_tensorflow}}, \code{\link[tensorflow]{tf_config}}, \code{\link[tensorflow]{tensorboard}}, \code{\link[tensorflow]{unique_dir}}, \code{\link[tensorflow]{run_dir}}}

  \item{tidyselect}{\code{\link[tidyselect]{starts_with}}, \code{\link[tidyselect]{ends_with}}, \code{\link[tidyselect]{contains}}, \code{\link[tidyselect]{matches}}, \code{\link[tidyselect]{everything}}, \code{\link[tidyselect]{one_of}}}
}}

