% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_options.R
\name{run_options}
\alias{run_options}
\title{Options for TensorFlow Routines}
\usage{
run_options(steps = 30L, model_dir = tf_setting("tflearn.model_dir",
  tempfile("tflearn_")), run_config = learn$RunConfig(tf_random_seed = 1))
}
\arguments{
\item{steps}{The number of steps to be used when running the associated model.}

\item{model_dir}{The location where model outputs should be written. Defaults
to a temporary directory within the \code{R} \code{\link{tempdir}}(), as
produced by \code{\link{tempfile}}().}
}
\description{
Used to control the various facets of layers constructed with the
modeling routines included in this package.
}
