% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_run_hooks.R
\name{session_run_hook}
\alias{session_run_hook}
\title{Create Session Run Hooks}
\usage{
session_run_hook(begin = NULL, after_create_session = NULL,
  before_run = NULL, after_run = NULL, end = NULL)
}
\arguments{
\item{begin}{\code{function()}: An \R function, to be called once before using the session.}

\item{after_create_session}{\code{function(session, coord)}: An \R function, to be called
once the new TensorFlow session has been created.}

\item{before_run}{\code{function(run_context)}: An \R function to be called before a run.}

\item{after_run}{\code{function(run_context, run_values)}: An \R function to be called
after a run.}

\item{end}{\code{function(session)}: An \R function to be called at the end of the session.}
}
\description{
Create a set of session run hooks, used to record information during
training of an estimator.
}
