% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_columns.R
\name{set_column_names}
\alias{set_column_names}
\alias{with_column_names}
\title{Set available feature column names}
\usage{
set_column_names(names)

with_column_names(names, expr)
}
\arguments{
\item{names}{Source of names (character vector, data frame, etc.)}

\item{expr}{Expression to evaluate with column names set.}
}
\value{
Previously set column names (invisibly)
}
\description{
Provide a list of names (or an R object with \code{names()} or \code{colnames()}) which
are valid for selection within \code{column_} feature column functions.
}
