% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_dnn_combined_estimators.R
\name{simple_linear_dnn_combined_recipe.default}
\alias{simple_linear_dnn_combined_recipe.default}
\title{#' @export
simple_linear_dnn_combined_recipe.formula <- function(x, data, ...) {
  parsed <- parse_formula(x)
  simple_linear_dnn_combined_recipe(data, parsed$response, parsed$features)
}}
\usage{
\method{simple_linear_dnn_combined_recipe}{default}(x, response,
  linear.features, dnn.features, run.config, ...)
}
\description{
#' @export
simple_linear_dnn_combined_recipe.formula <- function(x, data, ...) {
  parsed <- parse_formula(x)
  simple_linear_dnn_combined_recipe(data, parsed$response, parsed$features)
}
}

