% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_estimators.R
\name{svm_classifier}
\alias{svm_classifier}
\title{TensorFlow -- Support Vector Machines}
\usage{
svm_classifier(feature_columns, example_id_column, weight_column_name = NULL,
  model_dir = NULL, config = NULL)
}
\description{
Perform Support Vector Machines for binary classification using TensorFlow.
}
