% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{tf_linear_regression}
\alias{tf_linear_regression}
\title{TensorFlow -- Linear Regression}
\usage{
tf_linear_regression(recipe, run.options = run_options(), ...)
}
\arguments{
\item{recipe}{A recipe. See \code{\link{tf_recipe}} for more details.}

\item{...}{Optional arguments; currently unused.}

\item{tf.options}{Optional tools used for lower-level tuning of the TensorFlow
model fitting process. See \code{\link{tf_options}} for more information.}
}
\description{
Perform linear regression using TensorFlow.
}
\examples{
recipe <- tf_simple_recipe(mpg ~ drat, data = mtcars)
tf_linear_regression(recipe = recipe)
}

