% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_utils.R
\name{tf_simple_recipe}
\alias{tf_simple_recipe}
\title{Construct a Simple tflearn Recipe}
\usage{
tf_simple_recipe(data, response, features)
}
\arguments{
\item{data}{A tabular \R data set; typically a \code{data.frame}.}

\item{response}{The name of a variable within \code{data}, to be used
as the response.}

\item{features}{The names of variables within \code{data}, to be used
as features.}
}
\description{
Construct a simple recipe suitable for use with the higher-level
\code{tflearn} modeling routines. This recipe can be used to directly
model a response variable (\code{response}) as a function of a set of untransformed
features (\code{features}) in a dataset \code{data}.
}
\seealso{
Other recipes: \code{\link{tf_recipe}}
}

