% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tfestimators}
\alias{tfestimators-package}
\alias{tfestimators}
\title{High-level Estimator API in TensorFlow for R}
\description{
This library provides an R interface to the
\href{https://github.com/tensorflow/tensorflow/tree/r1.15/tensorflow/python/estimator}{Estimator}
API inside TensorFlow that's designed to streamline the process of creating,
evaluating, and deploying general machine learning and deep learning models.
}
\details{
\href{https://www.tensorflow.org}{TensorFlow} is an open source software library
for numerical computation using data flow graphs. Nodes in the graph
represent mathematical operations, while the graph edges represent the
multidimensional data arrays (tensors) communicated between them. The
flexible architecture allows you to deploy computation to one or more CPUs or
GPUs in a desktop, server, or mobile device with a single API.

The \href{https://www.tensorflow.org/versions/r1.15/api_docs/python/tf/all_symbols}{TensorFlow
API} is composed of a set of Python modules that enable constructing and
executing TensorFlow graphs. The tensorflow package provides access to the
complete TensorFlow API from within R.

For additional documentation on the tensorflow package see
\href{https://tensorflow.rstudio.com}{https://tensorflow.rstudio.com}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/tfestimators}
  \item Report bugs at \url{https://github.com/rstudio/tfestimators/issues}
}

}
\author{
\strong{Maintainer}: Tomasz Kalinowski \email{tomasz.kalinowski@rstudio.com}

Authors:
\itemize{
  \item JJ Allaire \email{jj@rstudio.com}
  \item Yuan Tang \email{terrytangyuan@gmail.com} (\href{https://orcid.org/0000-0001-5243-233X}{ORCID})
  \item Kevin Ushey \email{kevin@rstudio.com}
  \item Kevin Kuo \email{kevin.kuo@rstudio.com} (\href{https://orcid.org/0000-0001-7803-7901}{ORCID})
}

Other contributors:
\itemize{
  \item Daniel Falbel [contributor, copyright holder]
  \item RStudio [copyright holder, funder]
  \item  Google Inc. [copyright holder]
}

}
