% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_run_hooks_custom.R
\name{visualize_metrics}
\alias{visualize_metrics}
\title{Visualize the training or evaluation metrics}
\usage{
visualize_metrics(mode_key = c("train", "eval"))
}
\arguments{
\item{mode_key}{The mode when the metrics were collected that you want to visualize, e.g. "train"}
}
\value{
The directory used to save the metrics metadata and generated html file.
}
\description{
This function allows users to visualize the metrics produced in training or evaluation phases.
}
